+++ public/js/lib/utils/compatibility.js
@@ -1,9 +1,0 @@
-export function checkJavaScriptSupport() {
-	document.documentElement.classList.add('js');
-}
-
-export function hasModuleSupport() {
-	return ('supports' in HTMLScriptElement)
-		? HTMLScriptElement.supports('module')
-		: ('noModule' in document.createElement('script'));
-}